+++ test/refactor_nrepl/extract_definition_test.clj
@@ -1,8 +1,13 @@
 (ns refactor-nrepl.extract-definition-test
   (:require [clojure.java.io :as io]
-            [clojure.test :refer [deftest is]]
+            [clojure.test :refer [deftest is use-fixtures]]
+            [refactor-nrepl.core :as core]
             [refactor-nrepl.extract-definition :as sut]
             [refactor-nrepl.unreadable-files :refer [ignore-errors-str]]))
+
+(use-fixtures :once (fn [t]
+                      (binding [core/*skip-resources?* false]
+                        (t))))
 
 (defn- -extract-definition
   [name line col]