+++ test/single_segment_ns.clj
@@ -1,0 +1,3 @@
+(ns single-segment-ns)
+
+(defn foo [])