+++ test/refactor_nrepl/ns/class_search_test.clj
@@ -1,5 +1,6 @@
 (ns refactor-nrepl.ns.class-search-test
   (:require
+   [clojure.string :as string]
    [clojure.test :refer [deftest is]]
    [refactor-nrepl.ns.class-search :as sut]))
 
@@ -13,30 +14,50 @@
 (def non-initializable-classes
   '#{org.mozilla.javascript.SecureCaller})
 
+(defn handle [^Throwable e]
+  (is (or
+       ;; there are only ~5 in ~7922 classes that cause NoClassDefFoundError,
+       ;; see `#'acceptable-error-messages`.
+       ;; They don't have to do with classpath parsing so there's nothing to be fixed:
+       (contains? acceptable-error-messages (.getMessage e))
+       ;; Other internal classes introduced with JDK9:
+       (some (fn [prefix]
+               (-> e
+                   .getMessage
+                   (string/includes? prefix)))
+             ["org.graalvm"
+              "sun."
+              "jdk."]))
+      (-> e (.getMessage)))
+  e)
+
 (defn resolve-class [sym]
   (try
     (Class/forName (str sym)
                    false
                    (-> (Thread/currentThread) .getContextClassLoader))
     (catch NoClassDefFoundError e
-      ;; there are only ~5 in ~7922 classes that cause NoClassDefFoundError,
-      ;; see `#'acceptable-error-messages`.
-      ;; They don't have to do with classpath parsing so there's nothing to be fixed.
-      (is (contains? acceptable-error-messages (.getMessage e))
-          (-> e (.getMessage)))
-      e)
+      (handle e))
+    (catch ClassNotFoundException e
+      (handle e))
     (catch UnsupportedClassVersionError e
       e)))
 
 (defn result-can-be-ignored? [v]
   (or
    (instance? NoClassDefFoundError v)
+   (instance? ClassNotFoundException v)
    (instance? UnsupportedClassVersionError v)
    (contains? non-initializable-classes v)))
 
 (defn ok []
   (is (< 3000 (count @sut/available-classes))
       "There are plenty of completions offered / these's a test corpus")
+
+  (is (some #{'java.nio.channels.FileChannel} @sut/available-classes))
+  (is (some #{'java.io.File} @sut/available-classes))
+  (is (some #{'java.lang.Thread} @sut/available-classes))
+
   (is (< 3000 (count @sut/available-classes-by-last-segment)))
 
   (doseq [x @sut/available-classes